Update test cases and refactor test code
